fix: refactor wiki page component
